pyproject.toml
[tool.poetry]
name = "rekt-backend-api"
version = "0.1.0"
description = ""
authors = ["MansourRizk <mansour.rizk@hotmail.com>"]
readme = "README.md"
packages = [{ include = "rekt_backend_api" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.101.1"
uvicorn = "^0.23.2"
pydantic = { version = "^2.1.1", extras = ["email", "dotenv"] }

sqladmin = "^0.12.0"

asyncpg = "^0.27.0"
psycopg2-binary = "^2.9.6"
sqlalchemy = "^2.0.4"
alembic = "^1.10.2"

Unidecode = "^1.3"
boto3 = "^1.26.156"
boto3-type-annotations = "^0.3.1"
pyboto3 = "^1.13.18"

requests = "^2.31.0"
httpx = "^0.23.3"
python-multipart = "^0.0.5"
pyyaml = "^6.0"

redis = "^4.6.0"
fastapi-redis-cache = "^0.2.5"

pytest-mock = "^3.10.0"
pytest-github-report = "^0.0.1"
pytest-md = "^0.2.0"
pytest-emoji = "^0.2.0"
itsdangerous = "^2.1.2"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"


[tool.poetry.dev-dependencies]
black = "^23.7.0"
isort = "^5.10"
autoflake = "^2.2.0"
wemake-python-styleguide = "^0.18.0"
mypy = "^1.5.0"
flake8-fixme = "^1.1"
pytest = "^7.1"
pytest-cov = "^4.0"
pytest-asyncio = "^0.21.0"
pytest-env = "^0.8.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.2"

[tool.isort]
profile = "black"
src_paths = ["app"]
combine_as_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
